% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loon_ccmaps.R
\name{assign_levels}
\alias{assign_levels}
\title{Renames the levels for a vector of factors levels from '1' up to '3'
  depending on the number of existing levels in use}
\usage{
assign_levels(x)
}
\description{
Renames the levels for a vector of factors levels from '1' up to '3'
  depending on the number of existing levels in use
}
