% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loon_ccmaps.R
\name{assignLevels}
\alias{assignLevels}
\title{Renames the levels for a vector of factors levels from '1' up to '3'
  depending on the number of existing levels in use}
\usage{
assignLevels(x)
}
\description{
Renames the levels for a vector of factors levels from '1' up to '3'
  depending on the number of existing levels in use
}
