% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bindings.R
\name{bind_map2scat}
\alias{bind_map2scat}
\title{Bind polygon layer selection to scatterplot points}
\usage{
bind_map2scat(s, m, m_base, mapping)
}
\description{
Bind polygon layer selection to scatterplot points
}
