% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bindings.R
\name{bind_map2scat_add}
\alias{bind_map2scat_add}
\title{Bind polygon layer selections to scatterplot points
  (allows for multiple polygons to be selected)}
\usage{
bind_map2scat_add(s, m, m_base, mapping)
}
\description{
Bind polygon layer selections to scatterplot points
  (allows for multiple polygons to be selected)
}
