% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bindings.R
\name{bind_scat2map}
\alias{bind_scat2map}
\title{Bind scatterplot points selection to polygon map layer}
\usage{
bind_scat2map(s, m, mapping, orig_colors)
}
\description{
Bind scatterplot points selection to polygon map layer
}
