% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bindings.R
\name{bind_scat2map_col}
\alias{bind_scat2map_col}
\title{Bind scatterplot colors to polygon map layer}
\usage{
bind_scat2map_col(s, m, plotorder_exp, ids)
}
\description{
Bind scatterplot colors to polygon map layer
}
