% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bindings.R
\name{bind_scat2map_sel}
\alias{bind_scat2map_sel}
\title{Bind scatterplot points selection to polygon map layer}
\usage{
bind_scat2map_sel(s, m, mapping, plotorder_exp, ids)
}
\description{
Bind scatterplot points selection to polygon map layer
}
