% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bindings.R
\name{bind_zoompandelta}
\alias{bind_zoompandelta}
\title{Binding zoom, pan and delta along x and/or y direction for a list of widgets}
\usage{
bind_zoompandelta(direction = c("x", "y", "xy"), ...)
}
\description{
Binding zoom, pan and delta along x and/or y direction for a list of widgets
}
