% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{canada_demographics}
\alias{canada_demographics}
\title{Example Dataset: Canadian Demographic Statistics}
\format{A data frame with 288 rows and the following 11 columns:
\describe{
  \item{id}{Census division id.}
  \item{population}{Population.}
  \item{pop_density}{Population density.}
  \item{pct_immigrants}{Immigrant population percentage.}
  \item{bachelor_above}{Percentage of population with Bachelor's degree education or above.}
  \item{median_earnings}{Median annual earnings.}
  \item{employment}{Number of employed individuals.}
  \item{employ_rate_15plus}{Employment rate, those aged 15 or older.}
  \item{employ_rate_25_54}{Employment rate, those aged between 25 and 54.}
  \item{pop_density_cuberoot}{Cube root of population density.}
  \item{population_cuberoot}{Cube root of population.}
}}
\source{
Chen, Xin. 2011. “Visual Analysis of Immigration in Canada Based on Conditioned Choropleth Maps.” December. University of Waterloo.
}
\usage{
canada_demographics
}
\description{
A dataset comparing demographic statistics of 288 Canadian census divisions
from the 2006 Census.
}
\note{
This is just an example data set.
}
\keyword{datasets}
