% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/l_ccmaps.R
\name{cut_interval}
\alias{cut_interval}
\title{Partitions x (numeric vector) into up to three intervals, with break points
  defined by the minimum, cut_lwr, cut_upr and maximum.
  Break points need not be unique}
\usage{
cut_interval(x, cut_lwr, cut_upr)
}
\description{
Partitions x (numeric vector) into up to three intervals, with break points
  defined by the minimum, cut_lwr, cut_upr and maximum.
  Break points need not be unique
}
