% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loon_ccmaps.R
\name{cutInterval}
\alias{cutInterval}
\title{Partitions x (numeric vector) into up to three intervals, with break points
  defined by the minimum, cut1, cut2 and maximum.
  Break points need not be unique}
\usage{
cutInterval(x, cut1, cut2)
}
\description{
Partitions x (numeric vector) into up to three intervals, with break points
  defined by the minimum, cut1, cut2 and maximum.
  Break points need not be unique
}
