% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bindings.R
\name{disable_zoompandelta}
\alias{disable_zoompandelta}
\title{Disable zooming, panning and delta in the x and/or y direction for a list of widgets}
\usage{
disable_zoompandelta(direction = c("x", "y", "xy"), ...)
}
\description{
Disable zooming, panning and delta in the x and/or y direction for a list of widgets
}
