% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{edPov}
\alias{edPov}
\title{Example Dataset: U.S. Education and Poverty Levels}
\format{A data frame with 51 rows and the following 5 columns:
\describe{
  \item{state}{Full state name.}
  \item{ed}{Education rate.}
  \item{pov}{Poverty rate.}
  \item{region}{Major U.S. region.}
  \item{StateAb}{Abbreviated state name.}
}}
\source{
micromap R package
}
\usage{
edPov
}
\description{
A dataset comparing education and poverty levels among the states. Washington
DC is treated as a separate data point.
}
\note{
This is just an example data set.
}
\keyword{datasets}
