% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/l_ccmaps.R
\name{l_ccmaps}
\alias{l_ccmaps}
\title{Conditioned choropleth maps}
\usage{
l_ccmaps(tt = tktoplevel(), var_inspector = TRUE, spdf, respvar,
  respvar.lab = NULL, cond1var, cond1var.lab = NULL, cond2var,
  cond2var.lab = NULL, respbreaks = 3, size = 10, seg1col = "blue",
  seg2col = "darkgrey", seg3col = "red", optimize = FALSE, otry = 10,
  title = "CCmaps")
}
\arguments{
\item{respvar}{character name of the response value variable}

\item{cond1var}{character name of the first conditioning variable (controls
panel assignment in the vertical direction)}

\item{cond2var}{character name of the second conditioning variable (controls
panel assignment in the horizontal direction)}

\item{respbreaks}{determines how the response data is divided into three groups.
Can either be the integer 3 or a numeric vector of four break points.
Defaults to 3, in which case the response values are divided into tertiles}

\item{optimize}{logical value indicating whether panel assignment should be optimized
for \eqn{R^2}. Defaults to FALSE, in which case the conditioning data is
divided into tertiles}

\item{otry}{integer (greater than 0) indicating number of values to try
for optimization (see above). Required if \code{optimize = TRUE}.
Defaults to 10. A higher \code{otry} value leads to more precise estimates
at the cost of longer computation time}

\item{title}{character string of the title of the map. Appears in the title bar
of the plot window. Defaults to "CCmaps"}

\item{data}{SpatialPolygonsDataFrame to hold polygon coordinates and attributes,
including values for variables used for determining map coloring and panel positioning}
}
\description{
2-way panel of maps for visualizing multivariate data analysis
}
\examples{
\dontrun{

## Example 1
## Get data
library(maptools)
library(rgdal)
columbus <- readOGR(system.file("shapes/columbus.shp", package = "maptools")[1], verbose = F)

## Plot
l_ccmaps(spdf = columbus,
         respvar = "CRIME", cond1var = "PLUMB", cond2var = "HOVAL",
         optimize = TRUE,
         title = "Columbus Residential Burglaries and Vehicle Thefts")


## Example 2
## Get data
library(maptools)
library(rgdal)
nc.sids <- readOGR(system.file("shapes/sids.shp", package = "maptools")[1], verbose = F)

proj4string(nc.sids) <- CRS("+proj=longlat +ellps=clrk66")
row.names(nc.sids) <- as.character(nc.sids$FIPSNO)

nc.sids$ft.SID74 <- sqrt(1000)*(sqrt(nc.sids$SID74/nc.sids$BIR74) + sqrt((nc.sids$SID74+1)/nc.sids$BIR74))

nc.sids$ft.NWBIR74 <- sqrt(1000)*(sqrt(nc.sids$NWBIR74/nc.sids$BIR74) + sqrt((nc.sids$NWBIR74+1)/nc.sids$BIR74))

## Plot
l_ccmaps(spdf = nc.sids,
         respvar = "SID79", cond1var = "BIR79", cond2var = "SID74",
         optimize = FALSE,
         title = "North Carolina SIDS Rates")

}

}
