% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loon_ccmaps.R
\name{optimize}
\alias{optimize}
\title{Computes the R^2 measures by trying a number of possible partitioning points
  on the two conditioning variables, and returns the results in a data.frame
  sorted in decreasing order by the R^2}
\usage{
optimize(otry, data, resp, cond1, cond2)
}
\description{
Computes the R^2 measures by trying a number of possible partitioning points
  on the two conditioning variables, and returns the results in a data.frame
  sorted in decreasing order by the R^2
}
