% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/l_ccmaps.R
\name{r2_calc}
\alias{r2_calc}
\title{Calculates R^2 given the study variable values and the panel assignment
  based on the conditioning variables}
\usage{
r2_calc(x, group)
}
\description{
Calculates R^2 given the study variable values and the panel assignment
  based on the conditioning variables
}
