% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/l_ccmaps.R
\name{r2_optimize}
\alias{r2_optimize}
\title{Computes the R^2 measures by trying a number of possible partitioning points
  on the two conditioning variables, and returns the results in a data.frame
  sorted in decreasing order by the R^2}
\usage{
r2_optimize(otry, resp, cond1, cond2)
}
\description{
Computes the R^2 measures by trying a number of possible partitioning points
  on the two conditioning variables, and returns the results in a data.frame
  sorted in decreasing order by the R^2
}
