% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loon_ccmaps.R
\name{r2calc}
\alias{r2calc}
\title{Calculates R^2 given the study variable values and the panel assignment
  based on the conditioning variables}
\usage{
r2calc(x, g)
}
\description{
Calculates R^2 given the study variable values and the panel assignment
  based on the conditioning variables
}
