% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{us_lungmort}
\alias{us_lungmort}
\title{Example Dataset: U.S. Lung and Bronchus Cancer Mortality}
\format{A data frame with 51 rows and the following 5 columns:
\describe{
  \item{state}{Full state name.}
  \item{state.abbr}{Abbreviated state name.}
  \item{lung_bronch_death}{Lung and bronchus cancer mortality rate.}
  \item{pm25}{Outdoor fine particulate matter (PM2.5) level.}
  \item{income}{Annual income per capita.}
}}
\source{
CDC U.S. Cancer Statistics (\url{https://nccd.cdc.gov/uscs/cancersbystateandregion.aspx}),
  CDC WONDER Portal (\url{https://wonder.cdc.gov/NASA-PM.html}), and
  U.S. Bureau of Economic Analysis (\url{https://www.bea.gov/iTable/index_regional.cfm})
}
\usage{
us_lungmort
}
\description{
A dataset comparing the lung and bronchus cancer mortality rates with
fine particulate matter concentration levels and income per capita, 2009-2011 average.
Washington DC is treated as a separate data point.
}
\note{
This is just an example data set.
}
\keyword{datasets}
