% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{USstates}
\alias{USstates}
\title{Example Dataset: U.S. States Polygons}
\format{The format is formal class 'SpatialPolygonsDataFrame' [package "sp"].}
\source{
micromap R package
}
\usage{
USstates
}
\description{
An object of class 'SpatialPolygonsDataFrame' that was created from a
shapefile of U.S. state borders. Washington DC is treated as a separate area/polygon.
}
\note{
This is just an example data set.
}
\keyword{datasets}
