% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/l_ccmaps.R
\name{assign_colors}
\alias{assign_colors}
\title{Determines the colors to use for coloring regions. Needed since there might
  not be three categories for the study variable (e.g. when the first and
  second break point are equal)}
\usage{
assign_colors(min, max, cut1, cut2)
}
\description{
Determines the colors to use for coloring regions. Needed since there might
  not be three categories for the study variable (e.g. when the first and
  second break point are equal)
}
