% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/l_ccmaps.R
\name{l_ccmaps}
\alias{l_ccmaps}
\title{Conditioned choropleth maps in loon}
\usage{
l_ccmaps(tt = tktoplevel(), cc_inspector = TRUE, title = "CCmaps", spdf,
  respvar, respvar.lab = NULL, cond1var, cond1var.lab = NULL, cond2var,
  cond2var.lab = NULL, respbreaks = 3, size = 10, seg1col = "blue",
  seg2col = "darkgrey", seg3col = "red", optimize = FALSE, otry = 10)
}
\arguments{
\item{tt}{Tk top level window. Defaults to a new window}

\item{cc_inspector}{Whether to draw the custom inspector for CCmaps, which
allows for variable selection, variable label update, font size adjustment
and option to optimize \eqn{R^2}. Defaults to TRUE}

\item{title}{Title of the map. Appears in the title bar of the toplevel window.
Defaults to "CCmaps"}

\item{spdf}{\code{SpatialPolygonsDataFrame} object to hold polygon coordinates
and attributes. It should contain all variables used in analysis}

\item{respvar}{Name of the response value variable}

\item{respvar.lab}{Label for response value variable for slider. Defaults to NULL,
in which case \code{respvar} is used}

\item{cond1var}{Name of the first conditioning variable (controls
panel assignment in the vertical direction)}

\item{cond2var}{Name of the second conditioning variable (controls
panel assignment in the horizontal direction)}

\item{respbreaks}{Determines how the response data is divided into three groups
for coloring scheme. It can either be the integer 3 or a numeric vector of
four break points. Defaults to 3, in which case the response values are divided
into tertiles}

\item{size}{Font size for model value labels and\eqn{R^2}. Defaults to size 10}

\item{seg1col}{Color of first interval of points by \code{respvar} value.
Defaults to 'blue'}

\item{seg2col}{Color of second interval of points by \code{respvar} value.
Defaults to 'darkgrey'}

\item{seg3col}{Color of third interval of points by \code{respvar} value.
Defaults to 'red'}

\item{optimize}{Logical value indicating whether panel assignment should be
optimized for \eqn{R^2}. Requires a long time to compute. Defaults to FALSE,
in which case the conditioning data is divided into tertiles}

\item{otry}{Integer (greater than 0) indicating number of values to try
for optimization (see above). Required if \code{optimize = TRUE}.
Defaults to 10. A higher \code{otry} value leads to more precise estimates
at the cost of longer computation time}

\item{cond1.lab}{Label for first conditional variable for slider. Defaults to NULL,
in which case \code{cond1} is used}

\item{cond2.lab}{Label for second conditional variable for slider. Defaults to NULL,
in which case \code{cond2} is used}
}
\description{
2-way panel of maps for visualizing multivariate data
}
\examples{
\dontrun{

### Example 1
## Get data
library(maptools)
library(rgdal)
columbus <- readOGR(system.file("shapes/columbus.shp", package = "maptools")[1], verbose = F)

## Plot
l_ccmaps(title = "Columbus Residential Burglaries and Vehicle Thefts",
         spdf = columbus,
         respvar = "CRIME", cond1var = "PLUMB", cond2var = "HOVAL",
         optimize = TRUE)


### Example 2
## Get data
library(maptools)
library(rgdal)
nc.sids <- readOGR(system.file("shapes/sids.shp", package = "maptools")[1], verbose = F)

proj4string(nc.sids) <- CRS("+proj=longlat +ellps=clrk66")
row.names(nc.sids) <- as.character(nc.sids$FIPSNO)

nc.sids$ft.SID74 <- sqrt(1000)*(sqrt(nc.sids$SID74/nc.sids$BIR74) + sqrt((nc.sids$SID74+1)/nc.sids$BIR74))

nc.sids$ft.NWBIR74 <- sqrt(1000)*(sqrt(nc.sids$NWBIR74/nc.sids$BIR74) + sqrt((nc.sids$NWBIR74+1)/nc.sids$BIR74))

## Plot
l_ccmaps(title = "North Carolina SIDS Rates",
         spdf = nc.sids,
         respvar = "SID79", cond1var = "BIR79", cond2var = "SID74",
         optimize = FALSE)

}

}
