% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/l_micromaps.R
\name{l_micromaps}
\alias{l_micromaps}
\title{Linked Micromaps in loon}
\usage{
l_micromaps(top = tktoplevel(), mm_inspector = TRUE, title = "Micromaps",
  map.label = "Map", lab.label = "Labels", spdf, grouping = NULL,
  n_groups = NULL, variables, num_optvars = NULL, spacing = c("equal",
  "max"), color = NULL, size = 6, linkingKey = NULL,
  linkingGroup = NULL, sync = c("pull", "push"), ...)
}
\arguments{
\item{top}{Tk top level window. Defaults to a new window}

\item{mm_inspector}{Whether to draw custom inspector for the display, which
allows for variable selection, variable label update, font size adjustment,
and setting grouping of points. Defaults to TRUE. Once created, the inspector
can only be closed when the main display window is closed}

\item{title}{Title for linked micromaps. Appears in the title bar of
the toplevel window}

\item{map.label}{Label for maps (rightmost) panel}

\item{lab.label}{Label for labels (leftmost) panel}

\item{spdf}{\code{SpatialPolygonsDataFrame} object to hold polygon coordinates
and attributes. It should contain statistics used for comparing between regions}

\item{grouping}{An optional character vector specifying how many points to
have per row (from top to bottom of display)}

\item{n_groups}{An optional integer specifying how many rows to have.
If both \code{grouping} and \code{n_groups} are provided, \code{grouping}
is applied}

\item{variables}{List specifying variables to plot. Variable required: \code{id.var}
and \code{grouping.var}. Optional variables: more variables to plot.
\code{id.var} specifies the names of points. The \code{grouping.var} is
used to decide the ordering of points on the map. All variables,
with the exception of \code{id.var}, should be a list with \code{name},
and optional \code{xlab} and optional \code{label} arguments.
For example, for a variable named 'ed' used for grouping, the \code{variables}
argument would look like:
\code{variables = list(grouping.var = list(name = "ed", xlab = "Percent", label = "\% with University Education"))}}

\item{num_optvars}{Number of possible optional variables for \code{mm_inspector}.
Defaults to NULL, in which case it is determined based on how many variables are
provided to \code{variables}}

\item{spacing}{Spacing scheme for points - either 'equal' or 'max'. The 'equal'
scheme spaces points out equally, while the 'max' scheme ensures the
same amount of spacing between points as the row with the largest
number of points}

\item{color}{Color scheme. Defaults to NULL, in which case colors are
generated using \code{loon::loon_palette}. Cannot be 'cornsilk' or 'magenta',
which are reserved colors}

\item{size}{Size of glyphs for scatterplots and label panel. Defaults to 6}

\item{linkingKey}{Linking mechanism in \code{loon}. Points with the same
linkingKey value are linked together. Defaults to NULL, in which case the
linkingKey values are index values starting from 0}

\item{linkingGroup}{Linking mechanism in \code{loon}. Displays with the same
linkingGroup are linked together. Defaults to NULL, in which case the
linkingGroup is "Micromaps"}

\item{sync}{Can be either 'pull' or 'push', determines whether the initial
synchronization should adapt to the linked states of other linked plots ("pull")
or whether it should overwrite the states of the other linked plot with its
own linked states ("push")}

\item{...}{Other optional named arguments to modify states of scatterplots}
}
\value{
An object of classes \code{loon_micromaps} and \code{loon}, containing
  the Tk toplevel window, \code{linkingGroup} value, \code{linkingKey} values,
  and the handles for the \code{loon} plot objects corresponding to
  the labels panel, scatterplot panel(s) and map panel in list form
}
\description{
Compares different statistics across geographical regions
}
\examples{
\dontrun{

data("USstates", package = "micromap")
data("edPov", package = "micromap")

USstates@data <- merge(USstates@data, edPov, by.x = 'ST', by.y = 'StateAb')

mm <- l_micromaps(lab.label = 'States',
                  spdf = USstates,
                  variables = list(id.var = 'ST_NAME',
                                   grouping.var = list(name = 'pov', xlab = 'Percent'),
                                   var2 = list(name = 'ed', xlab = 'Percent')),
                  spacing = 'max', sync = 'push',
                  itemLabel = as.character(USstates@data$ST_NAME),
                  showItemLabels = T)

}

}
