% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/l_micromaps.R
\name{l_micromaps}
\alias{l_micromaps}
\title{Micromaps in loon}
\usage{
l_micromaps(tt = tktoplevel(), var_inspector = TRUE, spdf, labels,
  grouping = 5, ord.var, ord.var.axis, ord.var.label = NULL, var2 = NULL,
  var2.axis = NULL, var2.label = NULL, map.label = "Map",
  lab.label = "Labels", title = "Micromaps")
}
\description{
Micromaps in loon
}
\examples{
\dontrun{

library(micromap)

data("USstates")
data("edPov")

USstates@data <- merge(USstates@data, edPov, by.x = 'ST', by.y = 'StateAb', sort = F)

grouping <- 8

ord.var <- 'pov'
ord.var.axis <- 'Percent'
ord.var.label <- 'Poverty Rate'

var2 <- 'ed'
var2.axis <- 'Percent'
var2.label <- 'Education Rate'

lab.label <- 'States'

states <- USstates@data$ST_NAME

mm <- l_micromaps(spdf = USstates, labels = states, grouping = grouping,
                  ord.var = ord.var, ord.var.axis = ord.var.axis, ord.var.label = ord.var.label,
                  var2 = var2, var2.axis = var2.axis, var2.label = var2.label,
                  lab.label = lab.label)

}
}
