% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/l_micromaps.R
\name{l_micromaps}
\alias{l_micromaps}
\title{Micromaps in loon}
\usage{
l_micromaps(tt = tktoplevel(), var_inspector = TRUE, spdf,
  grouping = NULL, n_groups = NULL, map.label = "Map",
  lab.label = "Labels", title = "Micromaps", variables = list(id.var =
  NULL, ...), color = NULL, size = 6, linkingKey = NULL,
  linkingGroup = NULL, sync = c("pull", "push"), ...)
}
\description{
Micromaps in loon
}
\examples{
\dontrun{

library(micromap)

data("USstates")
data("edPov")

USstates@data <- merge(USstates@data, edPov, by.x = 'ST', by.y = 'StateAb', sort = F)

mm <- l_micromaps(spdf = USstates,
                  lab.label = 'States',
                  variables = list(id.var = 'ST_NAME',
                                   grouping.var = list(name = 'pov', xlab = 'Percent'),
                                   var2 = list(name = 'ed', xlab = 'Percent')),
                  linkingGroup = 'micromaps', sync = 'push')

}

}
